+++ HISTORY.md
@@ -26,7 +26,7 @@
             * RGB(A) value lists
         - new fundamentally simplified "Colors" library based on first-class colors, lets you mix, modify and transition colors
     * Hyper-mutation for arbitrary (sub-) lists: the CHANGE BY command also accepts list values in its first input slot
-    * Added sigmoid function ('σ') to the dropdown of the monadic operators reporter
+    * Added sigmoid (σ) and derivative sigmoid (∂σ) functions to the dropdown of the monadic operators reporter
     * Minor typography enhancements
         - new "txt_width" extension, reports the width of the given text at the specified font size and optional stylings
         - "Writing and formatting" library: new "width of text" reporter for determining the width of a text given the specified size and stylings 
@@ -54,6 +54,9 @@
     * Just Bars library: Only plot non-zero values, avoid drawing a "dot" for zero
 * **Translation Updates:**
     * German, including translation of the "Writing and formatting" library blocks
+
+### 2025-06-10
+* blocks, threads: added derivative sigmoid function ('∂σ') to the dropdown of the monadic operators reporter
 
 ### 2025-06-03
 * extensions: added "tts_stop" extension, un-blocks all processes currently waiting for speech recognition and advances to the next block